backend/requirements.txt
fastapi
uvicorn
tensorflow
pillow
python-multipart
opencv-python
